requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy>=2.0.35
psycopg2-binary>=2.9.9
pydantic>=2.5.2
pydantic-settings==2.1.0
pydantic-extra-types==2.4.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
email-validator>=2.1.0
bcrypt==3.2.2